FabsComponent fixed
